refactor(validator): split Iroh DKG tests
